move more common node methods into generator template
